Move favorite secondary functions to the another module
